+++ Platforms/Discord/Processing/removequote.py
@@ -5,7 +5,7 @@ if TYPE_CHECKING:
 from Utils.Classes.discordquote import DiscordQuote
 from Utils.Classes.discordcommand import DiscordCommand
 from Utils.Classes.discordcommandcontext import DiscordCommandContext
-from Platforms.Discord.utils import getDiscordServerQuotes
+from Platforms.Discord.db import getDiscordServerQuotes
 
 async def removeQuote(cls:"PhaazebotDiscord", Command:DiscordCommand, CommandContext:DiscordCommandContext) -> dict:
