+++ Platforms/Web/Processing/Api/Discord/Commands/get.py
@@ -8,7 +8,7 @@ import discord
 from aiohttp.web import Response, Request
 from Utils.Classes.webrequestcontent import WebRequestContent
 from Platforms.Web.Processing.Api.errors import apiMissingData
-from Platforms.Discord.utils import getDiscordServerCommands, getDiscordServerCommandsAmount
+from Platforms.Discord.db import getDiscordServerCommands, getDiscordServerCommandsAmount
 from Utils.Classes.discordwebuserinfo import DiscordWebUserInfo
 from Platforms.Web.Processing.Api.errors import apiMissingAuthorisation
 from Platforms.Web.Processing.Api.Discord.errors import apiDiscordGuildUnknown, apiDiscordMemberNotFound, apiDiscordMissingPermission
