+++ Platforms/Web/Processing/Api/Discord/Configs/Blacklistedwords/delete.py
@@ -9,7 +9,7 @@ from aiohttp.web import Response, Request
 from Utils.Classes.webrequestcontent import WebRequestContent
 from Utils.Classes.discordwebuserinfo import DiscordWebUserInfo
 from Utils.Classes.discordblacklistedword import DiscordBlacklistedWord
-from Platforms.Discord.utils import getDiscordServerBlacklistedWords
+from Platforms.Discord.db import getDiscordServerBlacklistedWords
 from Platforms.Web.Processing.Api.errors import (
 	apiMissingAuthorisation,
 	apiMissingData
