+++ Platforms/Web/Processing/Api/Discord/Configs/Exceptionroles/delete.py
@@ -9,7 +9,7 @@ from aiohttp.web import Response, Request
 from Utils.Classes.webrequestcontent import WebRequestContent
 from Utils.Classes.discordwebuserinfo import DiscordWebUserInfo
 from Utils.Classes.discordwhitelistedrole import DiscordWhitelistedRole
-from Platforms.Discord.utils import getDiscordServerExceptionRoles
+from Platforms.Discord.db import getDiscordServerExceptionRoles
 from Platforms.Web.Processing.Api.errors import (
 	apiMissingAuthorisation,
 	apiMissingData
