+++ Platforms/Web/Processing/Api/Discord/Configs/Gameenabledchannels/delete.py
@@ -9,7 +9,7 @@ from aiohttp.web import Response, Request
 from Utils.Classes.webrequestcontent import WebRequestContent
 from Utils.Classes.discordwebuserinfo import DiscordWebUserInfo
 from Utils.Classes.discordgameenablededchannel import DiscordGameEnabledChannel
-from Platforms.Discord.utils import getDiscordServerGameEnabledChannels
+from Platforms.Discord.db import getDiscordServerGameEnabledChannels
 from Platforms.Web.Processing.Api.errors import (
 	apiMissingAuthorisation,
 	apiMissingData
