+++ Platforms/Web/Processing/Api/Discord/Configs/Normaldisabledchannels/delete.py
@@ -9,7 +9,7 @@ from aiohttp.web import Response, Request
 from Utils.Classes.webrequestcontent import WebRequestContent
 from Utils.Classes.discordwebuserinfo import DiscordWebUserInfo
 from Utils.Classes.discordnormaldisabledchannel import DiscordNormalDisabledChannel
-from Platforms.Discord.utils import getDiscordServerNormalDisabledChannels
+from Platforms.Discord.db import getDiscordServerNormalDisabledChannels
 from Platforms.Web.Processing.Api.errors import (
 	apiMissingAuthorisation,
 	apiMissingData
