+++ Platforms/Web/Processing/Api/Discord/Configs/Nsfwenabledchannels/delete.py
@@ -9,7 +9,7 @@ from aiohttp.web import Response, Request
 from Utils.Classes.webrequestcontent import WebRequestContent
 from Utils.Classes.discordwebuserinfo import DiscordWebUserInfo
 from Utils.Classes.discordnsfwenablededchannel import DiscordNsfwEnabledChannel
-from Platforms.Discord.utils import getDiscordServerNsfwEnabledChannels
+from Platforms.Discord.db import getDiscordServerNsfwEnabledChannels
 from Platforms.Web.Processing.Api.errors import (
 	apiMissingAuthorisation,
 	apiMissingData
