+++ Platforms/Web/Processing/Api/Discord/Configs/Regulardisabledchannels/delete.py
@@ -9,7 +9,7 @@ from aiohttp.web import Response, Request
 from Utils.Classes.webrequestcontent import WebRequestContent
 from Utils.Classes.discordwebuserinfo import DiscordWebUserInfo
 from Utils.Classes.discordregulardisabledchannel import DiscordRegularDisabledChannel
-from Platforms.Discord.utils import getDiscordServerRegularDisabledChannels
+from Platforms.Discord.db import getDiscordServerRegularDisabledChannels
 from Platforms.Web.Processing.Api.errors import (
 	apiMissingAuthorisation,
 	apiMissingData
