+++ Platforms/Web/Processing/Api/Discord/Configs/Whitelistedlinks/delete.py
@@ -9,7 +9,7 @@ from aiohttp.web import Response, Request
 from Utils.Classes.webrequestcontent import WebRequestContent
 from Utils.Classes.discordwebuserinfo import DiscordWebUserInfo
 from Utils.Classes.discordwhitelistedlink import DiscordWhitelistedLink
-from Platforms.Discord.utils import getDiscordServerWhitelistedLinks
+from Platforms.Discord.db import getDiscordServerWhitelistedLinks
 from Platforms.Web.Processing.Api.errors import (
 	apiMissingAuthorisation,
 	apiMissingData
