+++ Platforms/Web/Processing/Api/Discord/Configs/edit.py
@@ -11,7 +11,8 @@ from Utils.Classes.discordserversettings import DiscordServerSettings
 from Utils.Classes.discordwebuserinfo import DiscordWebUserInfo
 from Utils.Classes.undefined import UNDEFINED
 from Utils.dbutils import validateDBInput
-from Platforms.Discord.utils import getDiscordSeverSettings, getDiscordRoleFromString
+from Platforms.Discord.db import getDiscordSeverSettings
+from Platforms.Discord.utils import getDiscordRoleFromString
 from Platforms.Discord.blacklist import checkBlacklistPunishmentString
 from Platforms.Web.Processing.Api.errors import (
 	apiMissingData,
