+++ Platforms/Web/Processing/Api/Discord/Levels/Medals/delete.py
@@ -8,7 +8,7 @@ import discord
 from aiohttp.web import Response, Request
 from Utils.Classes.webrequestcontent import WebRequestContent
 from Platforms.Web.Processing.Api.errors import apiMissingData
-from Platforms.Discord.utils import getDiscordUsersMedals
+from Platforms.Discord.db import getDiscordUsersMedals
 from Utils.Classes.discordwebuserinfo import DiscordWebUserInfo
 from Utils.Classes.discordusermedal import DiscordUserMedal
 from Platforms.Web.Processing.Api.errors import apiMissingAuthorisation
