+++ Platforms/Web/Processing/Api/Discord/Levels/Medals/get.py
@@ -7,7 +7,7 @@ import json
 import discord
 from aiohttp.web import Response, Request
 from Utils.Classes.webrequestcontent import WebRequestContent
-from Platforms.Discord.utils import getDiscordUsersMedals, getDiscordUsersMedalAmount
+from Platforms.Discord.db import getDiscordUsersMedals, getDiscordUsersMedalAmount
 from Platforms.Web.Processing.Api.errors import apiMissingData
 from Platforms.Web.Processing.Api.Discord.errors import apiDiscordGuildUnknown
