+++ Platforms/Web/Processing/Api/Discord/Levels/get.py
@@ -7,7 +7,7 @@ import json
 import discord
 from aiohttp.web import Response, Request
 from Utils.Classes.webrequestcontent import WebRequestContent
-from Platforms.Discord.utils import getDiscordServerUsers, getDiscordServerUserAmount
+from Platforms.Discord.db import getDiscordServerUsers, getDiscordServerUserAmount
 from Platforms.Discord.levels import Calc as LevelCalc
 from Platforms.Web.Processing.Api.errors import apiMissingData
 from Platforms.Web.Processing.Api.Discord.errors import apiDiscordGuildUnknown
