+++ Platforms/Web/Processing/Api/Discord/Quotes/get.py
@@ -7,7 +7,7 @@ import json
 import discord
 from aiohttp.web import Response, Request
 from Utils.Classes.webrequestcontent import WebRequestContent
-from Platforms.Discord.utils import getDiscordServerQuotes, getDiscordServerQuotesAmount
+from Platforms.Discord.db import getDiscordServerQuotes, getDiscordServerQuotesAmount
 from Platforms.Web.Processing.Api.errors import apiMissingData
 from Platforms.Web.Processing.Api.Discord.errors import apiDiscordGuildUnknown
