+++ Platforms/Web/Processing/Api/Discord/Twitchalerts/get.py
@@ -7,7 +7,7 @@ import json
 import discord
 from aiohttp.web import Response, Request
 from Utils.Classes.webrequestcontent import WebRequestContent
-from Platforms.Discord.utils import getDiscordServerTwitchAlerts, getDiscordServerTwitchAlertsAmount
+from Platforms.Discord.db import getDiscordServerTwitchAlerts, getDiscordServerTwitchAlertsAmount
 from Platforms.Web.Processing.Api.errors import apiMissingData
 from Platforms.Web.Processing.Api.Discord.errors import apiDiscordGuildUnknown
